+++ testing/java/com/google/android/material/testapp/AndroidManifest.xml
@@ -100,6 +100,11 @@
       android:name=".TextInputLayoutActivity"
       android:theme="@style/Theme.MaterialComponents.NoActionBar"/>
 
+    <activity
+      android:name=".TextInputLayoutWithIconsActivity"
+      android:theme="@style/Theme.MaterialComponents.NoActionBar"
+      android:exported="false"/>
+
 
   </application>
 </manifest>
