+++ testing/java/com/google/android/material/testapp/TextInputLayoutWithIconsActivity.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2019 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.android.material.testapp;
+
+import com.google.android.material.testapp.base.BaseTestActivity;
+
+/** Test activity that has TextInputLayouts with different supported icons. */
+public class TextInputLayoutWithIconsActivity extends BaseTestActivity {
+  @Override
+  protected int getContentViewLayoutResId() {
+    return R.layout.design_text_input_icons;
+  }
+}
