+++ testing/java/com/google/android/material/testapp/res/values/strings.xml
@@ -48,5 +48,5 @@
   <string name="textinput_hint">Hint to the user</string>
   <string name="textinput_helper">Helper text to the user</string>
   <string name="textinput_helper_not_enabled">Helper text to the user, helper not enabled</string>
-
+  <string name="textinput_custom_end_icon">Custom end icon</string>
 </resources>
